fix: re-wrap text and re-center text using mutations
