+++ src/locale/US/mutations/pala_multi_gtz.js
@@ -1,4 +1,5 @@
 import Logo from '../logos';
+import { altContentMediaQuery } from './mediaQueries';
 
 export default {
     'layout:text': [
@@ -35,7 +36,10 @@ export default {
         [
             'logo.type:alternative',
             ({ textSize }) => ({
-                styles: [`.message__logo-container { width: ${textSize * 9}px }`],
+                styles: [
+                    altContentMediaQuery(textSize * 23.8),
+                    `.message__logo-container { width: ${textSize * 9}px }`
+                ],
                 messageWidth: false,
                 logo: Logo.ALTERNATIVE.COLOR
             })
