+++ src/locale/US/mutations/pala_single_gtz.js
@@ -1,4 +1,5 @@
 import Logo from '../logos';
+import { altContentMediaQuery } from './mediaQueries';
 
 export default {
     'layout:text': [
@@ -29,7 +30,10 @@ export default {
         [
             'logo.type:alternative',
             ({ textSize }) => ({
-                styles: [`.message__logo-container { width: ${textSize * 9}px }`],
+                styles: [
+                    altContentMediaQuery(textSize * 29.3),
+                    `.message__logo-container { width: ${textSize * 9}px }`
+                ],
                 messageWidth: false,
                 logo: Logo.ALTERNATIVE.COLOR
             })
